# orders/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "AAW Marketplace API",
      "version": "1.0.0",
      "description": "Dokumentasi API untuk Cart dan Order"
    },
    "host": "localhost:8004",
    "basePath": "/api",
    "schemes": ["http"],
    "paths": {
      "/cart": {
        "get": {
          "summary": "Mendapatkan semua item di keranjang pengguna",
          "responses": {
            "200": { "description": "Daftar item di keranjang berhasil diambil" }
          }
        },
        "post": {
          "summary": "Menambahkan item ke keranjang",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "product_id": { "type": "string" },
                  "quantity": { "type": "integer" }
                }
              }
            }
          ],
          "responses": {
            "201": { "description": "Item berhasil ditambahkan ke keranjang" }
          }
        }
      },
...
      "/order": {
        "get": {
          "summary": "Mendapatkan semua pesanan pengguna",
          "responses": {
            "200": { "description": "Daftar pesanan berhasil diambil" }
          }
        },
        "post": {
          "summary": "Membuat pesanan baru",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "shipping_provider": { "type": "string" }
                }
              }
            }
          ],
          "responses": {
            "201": { "description": "Pesanan berhasil dibuat" }
          }
        }
      },
...
    },
    "securityDefinitions": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Please type Bearer before adding your token",
        "x-bearerInfoFunc": "bearer_auth"
      }
    },
    "security": [
        {
        "bearerAuth": []
        }
    ]
  }